requirements.txt
alembic==1.18.4
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
gunicorn==21.2.0
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.2.0
psycopg2-binary==2.9.9
PyJWT==2.12.0
python-dotenv==1.2.2
qrcode==8.2
requests==2.33.1
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6